pyproject.toml
[project]
name = "animeface-gan"
version = "0.1.0"
description = "DCGAN training on AnimeFace128 with Hydra, Lightning, and wandb."
authors = [{ name = "AnimeFace Team" }]
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "addict>=2.4.0",
    "datasets==2.19.1",
    "einops>=0.8.1",
    "hydra-colorlog>=1.2.0",
    "hydra-core>=1.3.2",
    "jaxtyping>=0.3.3",
    "modelscope>=1.33.0",
    "pillow>=12.0.0",
    "pytorch-lightning>=2.6.0",
    "rich>=14.2.0",
    "torch>=2.9.1",
    "torch-fidelity>=0.3.0",
    "torchvision>=0.24.1",
    "tqdm>=4.67.1",
    "typeguard>=4.4.4",
    "wandb>=0.23.1",
]

[project.optional-dependencies]
dev = [
    "pre-commit>=4.5.0",
    "ruff==0.6.9",
]

[dependency-groups]
dev = [
    "pre-commit>=4.5.0",
    "ruff==0.6.9",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/anime_gan"]

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.ruff.lint]
select = [
    "E",
    "F",
    "I",
    "B",
    "UP",
    "N",
]
ignore = ["F722"]

[tool.ruff.format]
quote-style = "double"
indent-style = "space"
line-ending = "lf"
